requirements.txt
tensorflow-gpu==2.5
dill
pandas
python-graphviz
pydot
pillow
huggingface_hub